5 acceptance testing

+++ app/controllers/questions_controller.rb
@@ -1,13 +1,18 @@
 class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: %i[index show]
 
-  expose :question,  -> { params[:question] ? user.questions.new(question_params) : Question.new }
-  expose :questions, -> { Question.all }
-  expose :answers,   -> { question.answers }
-  expose :user,      -> { current_user }
+  expose :questions,       ->{ Question.all }
+  expose :user,            ->{ current_user }
+  expose :answer,   build: ->(answer_params){ user.answers.new(answer_params) }, id: :answer_id
+  expose :question, build: ->(question_params){ user.questions.new(question_params) }
 
   def create
-    question.save ? redirect_to(questions_path) : render(:index)
+    if question.save
+      flash[:success] = 'Вопрос успешно создан'
+      redirect_to questions_path
+    else
+      render :index
+    end
   end
 
   def update
@@ -15,9 +20,17 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if user&.author_of?(question)
+      question.destroy
+      flash[:success] = 'Вопрос успешно удалён'
+    else
+      flash[:danger] = 'Вы не можете удалить чужой вопрос'
+    end
+
     redirect_to questions_path
   end
+
+  private
 
   def question_params
     params.require(:question).permit(:title, :body)
